# swagger/swagger-test.json
{
  "swagger": "2.0",
  "info": {
    "title": "N빵 API",
    "description": "SWM-ChocoBread N빵 API 명세서",
    "version": "1.0.0"
  },
  "host": "localhost:5005",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "http",
      "scheme": "bearer",
      "in": "header",
      "bearerFormat": "JWT"
    }
  },
  "paths": {
...
    "/deals/{dealId}/users/{userId}": {
      "get": {
        "tags": [
          "Deals"
        ],
        "summary": "거래 유저 상태(참여자, 제안자, 참여하지 않음)",
        "description": "",
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
...
    "/users/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "유저 정보 반환",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "닉네임 변경",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nick": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
...
  }
}